+++ Osiris/GUI.cpp
@@ -65,7 +65,7 @@ void GUI::renderAimbotWindow() noexcept
         ImGui::SetNextWindowSize(ImVec2(320.0f, 170.0f));
         ImGui::Begin("Aimbot", &window.aimbot, ImGuiWindowFlags_NoCollapse | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoScrollbar | ImGuiWindowFlags_NoScrollWithMouse);
         ImGui::Checkbox("Enabled", &config.aimbot.enabled);
-        ImGui::SliderFloat("Fov", &config.aimbot.fov, 0.0f, 180.0f);
+        ImGui::SliderInt("Fov", &config.aimbot.fov, 0, 180);
         ImGui::End();
     }
 }
